Add doc input `Range::class`.
