+++ docs/README.md
@@ -66,6 +66,7 @@ This is particularly useful for creating forms and other interactive elements on
 - [password](/docs/form/input/Password.md)
 - [radio](/docs/form/input/Radio.md)
 - [radio-list](/docs/form/input/RadioList.md)
+- [range](/docs/form/input/Range.md)
 - [text](/docs/form/input/Text.md)
 - [time](/docs/form/input/Time.md)
 - [url](/docs/form/input/Url.md)
