+++ tests/Input/Range/AttributeTest.php
@@ -153,26 +153,6 @@ final class AttributeTest extends TestCase
         );
     }
 
-    public function testMax(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="range-6582f2d099e8b" type="range" max="value">
-            HTML,
-            Range::widget()->id('range-6582f2d099e8b')->max('value')->render()
-        );
-    }
-
-    public function testMin(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="range-6582f2d099e8b" type="range" min="value">
-            HTML,
-            Range::widget()->id('range-6582f2d099e8b')->min('value')->render()
-        );
-    }
-
     public function testName(): void
     {
         Assert::equalsWithoutLE(
@@ -190,26 +170,6 @@ final class AttributeTest extends TestCase
             <input id="range-6582f2d099e8b" type="range" readonly>
             HTML,
             Range::widget()->id('range-6582f2d099e8b')->readonly()->render()
-        );
-    }
-
-    public function testRender(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="range-6582f2d099e8b" type="range">
-            HTML,
-            Range::widget()->id('range-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testStep(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="range-6582f2d099e8b" type="range" step="1">
-            HTML,
-            Range::widget()->id('range-6582f2d099e8b')->step(1)->render()
         );
     }
