+++ tests/Input/Range/CustomMethodTest.php
@@ -101,6 +101,16 @@ final class CustomMethodTest extends TestCase
         );
     }
 
+    public function testRender(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <input id="range-6582f2d099e8b" type="range">
+            HTML,
+            Range::widget()->id('range-6582f2d099e8b')->render()
+        );
+    }
+
     public function testSuffix(): void
     {
         Assert::equalsWithoutLE(
