+++ tests/Input/Range/ValidatorTest.php
@@ -1,0 +1,45 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PHPForge\Html\Tests\Input\Range;
+
+use PHPForge\Html\Input\Range;
+use PHPForge\Support\Assert;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @psalm-suppress PropertyNotSetInConstructor
+ */
+final class ValidatorTest extends TestCase
+{
+    public function testMax(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <input id="range-6582f2d099e8b" type="range" max="value">
+            HTML,
+            Range::widget()->id('range-6582f2d099e8b')->max('value')->render()
+        );
+    }
+
+    public function testMin(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <input id="range-6582f2d099e8b" type="range" min="value">
+            HTML,
+            Range::widget()->id('range-6582f2d099e8b')->min('value')->render()
+        );
+    }
+
+    public function testStep(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <input id="range-6582f2d099e8b" type="range" step="1">
+            HTML,
+            Range::widget()->id('range-6582f2d099e8b')->step(1)->render()
+        );
+    }
+}
